# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Zero-knowledge, self-hosted password vault API",
        "title": "vaultctl API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/vaults": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all vaults the authenticated user has access to",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vaults"
                ],
                "summary": "List vaults",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_presenters_api.VaultResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new vault with an encrypted vault key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vaults"
                ],
                "summary": "Create vault",
                "parameters": [
                    {
                        "description": "Vault creation payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.VaultCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.VaultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/vaults/{vaultId}/folders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List all folders in the specified vault",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Folders"
                ],
                "summary": "List folders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vault ID",
                        "name": "vaultId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_presenters_api.FolderResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new encrypted folder in the specified vault",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Folders"
                ],
                "summary": "Create folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vault ID",
                        "name": "vaultId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Folder payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.FolderCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.FolderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
        "/vaults/{vaultId}/folders/{folderId}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the encrypted name of a folder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Folders"
                ],
                "summary": "Rename folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vault ID",
                        "name": "vaultId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Folder ID",
                        "name": "folderId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New encrypted name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.FolderCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.FolderResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a folder (items in the folder are not deleted, just unlinked)",
                "tags": [
                    "Folders"
                ],
                "summary": "Delete folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vault ID",
                        "name": "vaultId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Folder ID",
                        "name": "folderId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
...
        "/vaults/{vaultId}/members": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new member to a shared vault with an encrypted vault key",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Sharing"
                ],
                "summary": "Share vault",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vault ID",
                        "name": "vaultId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Sharing payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ShareVaultRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_presenters_api.ErrorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_presenters_api.ErrorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "field": {
                            "type": "string"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "internal_presenters_api.VaultResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "encryptedVaultKey": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "senderId": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "wrapSignature": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\" (include the word Bearer)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}